More conversions to long.

+++ FloppyControlApp.net7/WindowsFormsApplication2/WindowsFormsApplication2/FloppyControlErrorCorrection.cs
@@ -317,16 +317,16 @@ namespace FloppyControlApp
 
             if (!BSBlueSectormapRadio.Checked) // there's no relevant data when this radio button is checked
             {
-				int SectorViewEnd = 0;
-				int SectorHeaderOffset;
-				int scatoffset = 0;
-				int scatlength = 100;
+				long SectorViewEnd = 0;
+				long SectorHeaderOffset;
+				long scatoffset = 0;
+				long scatlength = 100;
 
 				// Zoomed out
 				if (ScatterMinTrackBar.Value == 0 && ScatterMaxTrackBar.Value == 0)
 				{
 					SectorHeaderOffset = processing.sectordata2[processing.sectordata2[indexS1].HeaderIndex].rxbufMarkerPositions;
-					if (processing.sectordata2.Count > indexS1 + 1) SectorViewEnd = processing.sectordata2[indexS1 + 1].rxbufMarkerPositions;
+					if (processing.sectordata2.Count > indexS1 + 1) SectorViewEnd = processing.sectordata2[indexS1 + 1L].rxbufMarkerPositions;
 
 					if (SectorViewEnd - SectorHeaderOffset > 10000) SectorViewEnd = SectorHeaderOffset + 10000;
 					 scatoffset = SectorHeaderOffset + (int)ScatterMinTrackBar.Value + (int)ScatterOffsetTrackBar.Value;
